[exporterhelper] Remove metrics observer adapter
Use the same queue observer interface internally and publicly so ObsMetrics can pass directly into exporterhelper.
